autoapply-x/requirements.txt
# autoapply-x dependencies

# Web scraping and automation
playwright==1.41.0
beautifulsoup4==4.12.2
httpx==0.25.2

# Configuration handling
pyyaml==6.0
python-dotenv==1.0.0

# Utilities
asyncio==3.4.3
tqdm==4.66.1

# For LLM API integration
openai==1.3.0  # Compatible with OpenRouter API
requests==2.31.0
openrouter==0.3.0  # Optional OpenRouter client library

# For NLP processing (optional for future features)
mistralai==0.0.7
